% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpart.R
\name{devpart}
\alias{devpart}
\title{Deviance partitioning for three groups of predictors.}
\usage{
devpart(model, A, B, C, labels = c("A", "B", "C"), sep = "&")
}
\arguments{
\item{model}{Model object with deviance and null.deviance names.}

\item{A}{Character vector with names of the predictors of the first group.}

\item{B}{Character vector with names of the predictors of the second group.}

\item{C}{Character vector with names of the predictors of the third group.}

\item{labels}{Character vector with group labels. The default labels are "A", "B", and "C".}

\item{sep}{Character used to construct labels for different group intersections.}
}
\value{
Numeric vector with the following 8 elements:
1. Deviance only explained by A
2. Deviance only explained by B
3. Deviance only explained by C
4. Deviance not explained by C, but not allowing being attributed unambiguously to A or B
5. Deviance not explained by B, but not allowing being attributed unambiguously to A or C
6. Deviance not explained by A, but not allowing being attributed unambiguously to B or C
7. Deviance explained by A or B or C, without allowing being attributed unambiguously to any of them
8. Unexplained part of the deviance

The elements are named based on the labels: "A", "B", "C", "A&B", "A&C", "B&C", "A&B&C", "Unexplained".
}
\description{
Deviance partitioning for three groups of predictors.
}
