% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpart.R
\name{plot_venns}
\alias{plot_venns}
\title{Plot the Venn's/Euler's diagramms.}
\usage{
plot_venns(
  venns,
  labels = c("A", "B", "C"),
  cols = c("red", "green", "blue"),
  expand = c(0.1, 0.1),
  hjust = NULL,
  vjust = NULL,
  unexpl.position = c(1, 0),
  alpha = 0.4
)
}
\arguments{
\item{venns}{A list of two numeric vectors as returned by the function get_venns.}

\item{labels}{Character vector of group labels thaht will appear as labels in the plot.}

\item{cols}{Character vector specifying the colors used to fill the circles.}

\item{expand}{Numeric vector of length two, with a fraction by which the plot window should be extended in x and y directions in order to include all the labels. Default is 0.1 and 0.1 (i.e. 10% expansion in both directions).}

\item{hjust}{Numeric vector of length three, specifying the horizonthal adjustments of the group labels (i.e. the relative positions of the anchor points of the labels, with 0 ... anchor point is at the left edge of the label, 1 ... anchor point is at the right edge of the label). Values can exceed the (0, 1) interval.}

\item{vjust}{Numeric vector of length three, specifying the vertical adjustments of the group labels (i.e. the relative positions of the anchor points of the labels, with 0 ... anchor point is at bottom edge of the label, 1 ... anchor point is at the top edge of the label). Values can exceed the (0, 1) interval.}

\item{unexpl.position}{Character of length two, specifying the relative position of the label for the unexplained part of the deviance in the plot, in x and y direction. (0, 0) means lower left corner, (1, 1) means upper right corner. Values can exceed the (0, 1) interval.}

\item{alpha}{Numeric value between 0 and 1 specifying the transparency of the circles. The default is 0.4.}
}
\value{
A ggplot object visualizing the input deviance partitioning using three intersecting circles with areas of all parts optimized to reflect the deviance partitioning.
}
\description{
Plot the Venn's/Euler's diagramms.
}
