% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpart.R
\name{rsq}
\alias{rsq}
\title{Explained deviance of a model using the simple McFadden formula.}
\usage{
rsq(model)
}
\arguments{
\item{model}{Model object with deviance and null.deviance names.}
}
\value{
The fraction of the model's null deviance explained by all model predictors.
}
\description{
Explained deviance of a model using the simple McFadden formula.
}
